api/requirements.txt
asgiref==3.8.1
boto3==1.34.125
botocore==1.34.125
Django==5.0.6
django-admin-autocomplete-filter==0.7.1
django-admin-autocomplete-list-filter==1.0.1
django-admin-json-editor==0.2.3
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
mssql-django==1.4.2
PyJWT==1.7.1
pyodbc==5.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1